backend/pyproject.toml
[project]
name = "aiphotokey-backend"
version = "0.1.0"
description = "FastAPI backend for AIPhotoKey"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "pillow>=11.0.0,<12.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "rawpy>=0.25.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
ai = [
  "ultralytics>=8.3.0,<9.0.0",
]
caption = [
  "transformers>=4.49.0,<5.0.0",
]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
